gather-shop/requirements.txt
fastapi>=0.115,<1
uvicorn>=0.30,<1
pydantic>=2,<3
httpx>=0.27,<1
python-dotenv>=1,<2
